pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "boio"
version = "0.1.0"
description = "An opinionated way of doing AI - Task-oriented ai nodes for general-computing"
readme = "README.md"
authors = [{name = "SURUS AI", email = "contact@surus.lat"}]
license = "MIT"
requires-python = ">=3.8"
keywords = ["ai", "transcription", "audio", "whisper", "canary", "speech-to-text"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.25.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://surus.lat"
Repository = "https://github.com/surus-lat/boio"
Documentation = "https://docs.surus.lat"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "black>=22.0.0",
    "ruff>=0.1.0",
    "twine>=6.1.0",
    "build>=1.2.2.post1",
]

audio = [
    "pydub>=0.25.0",
]

[tool.setuptools.packages.find]
where = ["."]
